# backend-api/.build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GetUserResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string",
						"nullable": true
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string",
						"nullable": true
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetReviewResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string",
						"nullable": true
					},
					"idCaregiver": {
						"type": "number",
						"format": "double"
					},
					"patient": {
						"properties": {
							"urlImage": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"rating",
					"idCaregiver",
					"patient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					},
					"idCaregiver": {
						"type": "number",
						"format": "double"
					},
					"idPatient": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rating",
					"comment",
					"idCaregiver",
					"idPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPatientResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"properties": {
							"urlImage": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"cpf",
					"birthDate",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetCaregiverResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"corenNumber": {
						"type": "string"
					},
					"biography": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"properties": {
							"urlImage": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"corenNumber",
					"biography",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetAddressResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"street": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"complement": {
						"type": "string",
						"nullable": true
					},
					"neighborhood": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"street",
					"number",
					"neighborhood",
					"city",
					"state",
					"postalCode",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressRequest": {
				"properties": {
					"street": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"complement": {
						"type": "string",
						"nullable": true
					},
					"neighborhood": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"street",
					"number",
					"neighborhood",
					"city",
					"state",
					"postalCode",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"firebase": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "homecare-backend-api",
		"version": "1.0.0",
		"description": "API para o sistema de agendamento de atendimentos HomeCareApp",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Lista de usuários e seus dados",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetUserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Retorna todos os usuários cadastrados",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria um novo usuário",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRequest"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Dados do usuário",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um usuário pelo seu ID",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Altera os dados de um usuário",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Remove o registro de um usuário da base",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/firebase-user": {
			"get": {
				"operationId": "GetUserByFirebaseUid",
				"responses": {
					"200": {
						"description": "Dados do usuário",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um usuário pelo firebase UID dele",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": []
			}
		},
		"/reviews": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Lista de todas as avaliações",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetReviewResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Busca a lista de as avaliações na base",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"description": "Filtra pelo ID do cuidador (Somente um filtro deve ser usado por consulta)",
						"in": "query",
						"name": "idCaregiver",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "idPatient",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Retorna o ID da avaliação criada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria uma nova avaliação",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Retorna a avaliação consultada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetReviewResponse"
								}
							}
						}
					}
				},
				"summary": "Busca de uma avaliação pelo seu ID",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Atualiza uma avaliação",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Deleta a avaliação",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/patients/user/{idUser}": {
			"get": {
				"operationId": "GetPatientByUserId",
				"responses": {
					"200": {
						"description": "Retorna os dados do paciente consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPatientResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um paciente específico pelo ID do usuário associado",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUser",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/care-professionals/user/{idUser}": {
			"get": {
				"operationId": "GetCareProfessinalByUserId",
				"responses": {
					"200": {
						"description": "Retorna os dados do cuidador consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCaregiverResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um cuidador específico pelo ID do usuário associado",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUser",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/addresses": {
			"get": {
				"operationId": "GetAllAddresses",
				"responses": {
					"200": {
						"description": "Retorna todos os endereços e seus dados",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetAddressResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Busca por todos os endereços cadastrados na base",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAddress",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria um novo endereço",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest"
							}
						}
					}
				}
			}
		},
		"/addresses/{id}": {
			"get": {
				"operationId": "GetAddressById",
				"responses": {
					"200": {
						"description": "Exibe os dados do endereço",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAddressResponse"
								}
							}
						}
					}
				},
				"summary": "Busca por um endereço pelo seu ID",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"summary": "Atualiza um endereço pelo ID",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveAddress",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"summary": "Remove o registro de um endereço da base",
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}