# backend-api/src/models/address/dtos/GetAddressResponse.ts
export interface GetAddressResponse {
  id: number;
  street: string;
  number: string;
  complement?: string | null;
  neighborhood: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
}